Fix floating-point overflow for binary, octal, and hex literals.
Summary: Literals that are hex, oct, or binary are promoted to doubles when
they set the sign bit. We were using strtoull instead of strtoll.

We also weren't calling zend_oct_strtod or zend_bin_strtod (we didn't
even have it, I pulled it in from php-src) in
ScalarExpression::getVariant().

Reviewed By: @ptarjan

Differential Revision: D1409285