remove travis from readme
Summary: Travis hasn't been very high signal for us. It can't compile things in time, and we can't seem to keep all the tests passing on it. Our internal CI works great, and the move to open phabricator will expose this.

I'll leave it on for PRs, but I think showing the badge so priminantly is mis-selling out much we actually test this thing.

Reviewed By: @JoelMarcey

Differential Revision: D1409835