Clean up uses of Type::unbox() in MInstrTranslator
Summary: We were assuming that the result of a LdRef instruction would always
be typeParam.unbox(), but this isn't the case. I went through all the uses in
minstrtranslator and audited each one, either commenting on why it's ok or
fixing it to use ldRefReturn. The call to checkMIState() is now after
emitBaseOp() because none of the base ops actually needed it and this allows us
to use m_base in checkMIState() instead of duplicating some of its logic.

Reviewed By: @jdelong

Differential Revision: D1410177