Fix double-inlining bug w/ singleton inlining
Summary: My singleton inlining machinery assumed that the regular inliner could
never inline singleton patterns.  If this is false, we'll sometimes
inline a function twice, which appears to be happening in practice in
perflab.

I'll try to move the singleton stuff into the new inlining machinery
that I have diffs up for now.

Reviewed By: @bertmaher

Differential Revision: D1412015