Backtrace generation has been moved out of the execution context and into a new class, BackTrace
Summary: There are several diffs out that currently add arguments to the debugBacktrace function (D1405194, D1407595). With both the proposed diffs and the current arguments to debugBacktrace, the function will have around 10 optional arguments. To keep the code base readable it seems reasonable to create a class that handles the creation of backtraces. While there are a lot of lines of code in this diff, almost all changes are a direct copy of the old debugBacktrace function.

Reviewed By: @jdelong

Differential Revision: D1412812