Fix reoptimize() for control-flow
Summary: If optimizeInst() returns an SSATmp that is different from the
destination of the instruction, we add a "Mov" instruction to move
it to the correct destination. However, this was not correctly
handling the case when the optimized instruction has a taken branch.

This diff fixes it to add the Mov on the fall-through branch.

Reviewed By: @swtaarrs

Differential Revision: D1412984