Call $this->__init__() even in native classes
We (WMF) have many classes that look like:

<<__NativeData("ZendCompat")>>
class LuaSandboxError extends Exception {}

Native::nativeDataInstanceCtor() is mostly a duplicate of
ObjectData::newInstance(), but the Native function inappropriately omits
the __init__() call. So for the test case above, the exception trace is
not initialised.

Fix this, and also give a more informative error message when
Exception::__init__() is not called prior to
Exception::getTraceAsString().