Add literal table to MCGenerator
Summary: Adds a value->address map to mcgenerator for generating
large constants that we can point to using rip-relative
addressing. Use this mode when loading doubles into SIMD
registers on x64, and when we can't get a trampoline.

Also add a literal table to struct CodeGenFixups so we can
keep track of pending literals and only commit them to the
global table when we know we won't roll back.

Reviewed By: @markw65

Differential Revision: D1413464