hhi cleanup
Summary: Clean revert of the following three diffs:

{D1364920}
{D1368049}
{D1376096}

plus the NEWS change.

With those reverted, we now consider hhi embedding to be the default
and non-optional; OSS users should get a suitably useful error message
to help upgrade).

This isn't going to be the normal upgrade path for more involved
changes, but this one is easy enough and can be given a good enough
error message we're just going to do this.

Reviewed By: @elgenie

Differential Revision: D1412234