Fix bug renumbering region blocks
Summary: For Profile translations, the first block is renumbered to match the
TransID of the translation, so that we can easily map BlockIds to
TransIDs for these translations.  However, the arcs involving such
blocks were not being adjusted accordingly.  This diff fixes this, and
also add an assertion to make sure that we only process blocks in
non-topological order in translateRegion when loops are enabled.

Reviewed By: @edwinsmith

Differential Revision: D1411119