Implement a basic version of first-class enums in HHVM.
Summary: Implement a basic version of first-class enums in HHVM. Enums are basically represented as classes, but require special handling to use the underlying type for type checks when used as type hints.

Currently raises a fatal error if the underlying type can't be resolved when the Class is created (because it is a typedef or an enum that hasn't been defined yet). This is kind of unfortunate.

Reviewed By: @paroski

Differential Revision: D1454278