Fix more "newtype as" issues
Summary: @dlreeves found a few more places that we need to unwrap a
`newtype X as Y`. This is probably going to be whackamole for a little
while unfortunately. @pikatchu suggested I add the `Tabstract` case above
any `Tgeneric` case, but there's a //lot// of them, and I think this
approach is somewhat less messy.

Reviewed By: @pikatchu

Differential Revision: D1414279