removed a bad assert that could potentially fail
Summary: When an ArrayObject or ArrayIterator is cast to an array the convert_to_array function would grab the storage property of the array and return the value. If someone used reflection to make the storage property visible then unset the property an assert would be tripped in the convert_to_array function.

Reviewed By: @ptarjan

Differential Revision: D1419817