Fix autoloading for "type" and "newtype" statements
Summary: If the name on the right hand side of a "type" or "newtype" statement does
not name an existing class/interface/trait or an existing type alias, HHVM
tries both class-autoloading and type-autoloading. The problem is that if
class-autoloading doesn't find anything, HHVM calls the failure callback
before trying type-autoloading, which can cause problems for correct
programs (see the test cases).

This diff fixes HHVM to try both class-autoloading and type-autoloading
before calling the failure callback for "type" and "newtype" statements.

Reviewed By: @elgenie

Differential Revision: D1410570