Add MYSQLI_STMT_ATTR_UPDATE_MAX_LENGTH logic to store_result
Summary: This is the same as #3016 but moves the implementation to PHP instead.

It actually does not fix the bug itself because PHP behaves the same.
The cause of the segfault is that some people run out of memory resulting in a bad call to calloc.
calloc returns a nullptr which later in mysql is dereferenced.

PHP's solution to this is to call store_result before binding. But this will also fail in hhvm.
This fixes it.

Thanks SiebelsTim for writing the original diff in C++

TL;DR: TEXT in mysql is now variable length instead of 4GB when calling store_result

Fixes #2377
Fixes #2921

Reviewed By: @JoelMarcey

Differential Revision: D1407344