Add stats in hhbbc about persistent static properties
Summary: We're going to have to redo this in hhbbc if we want to turn off
inferTypes, so I was looking at how many there are.  (Right now there's also an
issue: if you turn off inferTypes, it appears that every static property gets
marked as persistent, because isIndirectAltered defaults false everywhere, and
we'll behave horribly wrongly at runtime.)

I think it would probably be pretty trivial to do this for private statics
right now, but a huge majority of the cases of this are public statics from
thrift classes, so it's probably necessary to make something that can track
generally which statics could be assigned to.

Reviewed By: @alexmalyshev

Differential Revision: D1439901