Unbreak DynamicYield in javascript
Summary: js_of_ocaml doesn't provide the Str library, so ages ago, somebody made a fake one so our existing code still worked.

Unfortunately, the javascript regexp library uses javascript's syntax for regexes, instead of ocaml syntax.

Because of this, DynamicYield has never worked in FBIDE. This used to be ok for typechecking, because we'd never bring in every dependency from Hack, so it would just assume these methods were defined in PHP. (So nobody ever caught it.) It's never worked for autocomplete (but I never realized that it *should* work for autocomplete)

There's no library method to convert regexes from ocaml to js format, so I did a terrible hack that makes this work for us. (I admit this is awful :( )

We're only using Str regex stuff for dynamic yield in javascript, so I was able to verify that this fixes it.

The problem for our regexes is that in ocaml, you escape ( and ) to indicate that they're a matching group, while in javascript, you do not.

Reviewed By: @jwatzman

Differential Revision: D1440297