Compute FPI PrepKinds using method names in hhbbc
Summary: This is one thing hphpc still does that hhbbc doesn't.  But the code
that sets Expression::RefValue into the context bits is incorrect if you don't
run the inferTypes pass, so we can't use it if we want to stop doing that.
Redoing the optimization here is pretty easy.

Reviewed By: @edwinsmith

Differential Revision: D1443303