unset another cached variable for dwarf detection
Summary: We're not actually broken right now because of the way we fall-back to
dwarf_producer_init_c, but a function exists check is still being
skipped due to CMake caching. That could potentially confuse things
if/when another init function were to be added in the future.
Closes #3200
Created from: https://github.com/facebook/hhvm/pull/3200

Reviewed By: @ptarjan

Differential Revision: D1438631