Allow negative TransIDs in certain contexts
Summary: For the new inlining stuff, we need to be able to convert negative
BlockIds into nontrivial ProfSrcKeys, so that we can correctly track the
set of instructions that need interpreting, but without changing the
behavior of anything else.

We're already using JIT::BlockIds and TransIDs in a very transparent
manner, so let's just give them the same types and logical ranges.

Reviewed By: @ottoni

Differential Revision: D1444737