Replace regexes with string functions in DynamicYield
Summary: As @elgenie suggested in D1440297, instead of my hacky fix, I've rewritten the regexes to not be regexes.

I also deleted str.ml completely. I do not know if we want to do this or not. This would prevent people from using Str regexes in any part of the codebase that is being compiled to javascript, so that nobody starts adding them without realizing that they don't really work in javascript.

Reviewed By: jwatzman, eletuchy

Differential Revision: D1444831