Share more code in FCall* opcodes, fix a bug in FCallArray and FCallUnpack
Summary: Neither FCallArray nor FCallUnpack were properly handling functions
that never return (the latter copied the bug from the former).  This wasn't a
problem 'in practice' for FCallArray because currently it never has a resolved
func on the ActRec right now (because FPushCuf doesn't try to figure out which
func you're calling), but it was possible to cause hhbbc to crash with an
unpacking call to a function that can't return.  Fix this while setting up the
code to share a little more of the logic.

This diff also makes it so FCall (non-D) with a known resolved func will try to
look up the return type.  This can happen with the new kind of name-only
function resolution from D1443303, but currently they don't ever produce return
types better than TInitGen, so it's a no-op for now.

Reviewed By: @elgenie

Differential Revision: D1448343