Support nested uses of reduce() in hhbbc
Summary: The implementation of reduce and impl claim you aren't allowed to nest
opcodes that do reduce calls, but the code was doing it anyway.  In practice we
were never hitting the case, but we do after turning off FPassC/V changes in
emitter.cpp, so now we should support it, which isn't too hard.

Reviewed By: @edwinsmith

Differential Revision: D1447745