Refactor ArgumentInfo and expose on the command line
Summary: Moved Argument info to use TypingHooks and made a command line API for it.

No promises on a stable API here yet, so I left the description empty so that lots of people don't start using it.

This is purely cleanup with regard to the logic. I haven't made any real changes.

Reviewed By: @jwatzman

Differential Revision: D1450096