Hide benign git error when running fbmake
Summary: hphp has a prerun command that runs before *every* fbmake. It runs git
to try to compute a hash for the hphp directory. Since their hash computation is
somewhat git specific, and would be difficult to port to Mercurial, let's just
hide the error that it produces for now.

Reviewed By: @sid0

Differential Revision: D1450699