Support "yield x => y" syntax
Summary: This syntax has existed for a while but the typechecker never
supported it. Let's do it. Mostly involves changing `Yield` to take an
`afield` instead of an expression, as well as a little bit of cleanup of
cruft from the days when we faked `async` with `yield` that makes no
sense nowadays and got in the way.

Reviewed By: @elgenie

Differential Revision: D1447853