Removed HOTPROFILER compile flag
Summary: This diff removes the HOTPROFILER flag.

I'm not sure if this is the best approach or if this should even be done at
all, but @swtaarrs said he thought this made sense so I figured I'd send this
out for review.

Currently the HOTPROFILER flags is always set, but the hooks into the
hotprofiler are still guarded by ifdefs. @swtaarrs said this was because it
used to have a non-zero overhead, but this is no longer the case. The
overhead would be in `event-hooks.cpp` but we don't run the hotprofiler code
unless event hooks have been enabled- and we have this check even
without the HOTPROFILER flag.

This matters for external profilers, namely xdebug. Since HOTPROFILER is
always set, this isn't actually required, but I'm not sure if it makes sense
for external profilers (implementations of `Profiler`) in general to
rely on the HOTPROFILER flag being set to work.

I just added anyone who commented on this in IRC as a reviewer. So feel free
to remove yourself if you'd like.

Reviewed By: @swtaarrs

Differential Revision: D1433761