Added xdebug_get_declared_vars
Summary: This diff adds the xdebug function xdebug_get_declared_vars.
xdebug_get_declared_vars returns an array containing the names of all the
variables in scope. It differs from the php function get_defined_vars at
the top level, because the environment variables are not included. Also
variable values are not included.

Depends on D1401578 in order to match php5 xdebug results in closures.

Reviewed By: @ptarjan

Differential Revision: D1401714