Retry tracelet selection if inlining fails
Summary: Apparently, if we have a nested callstack a() -> b() -> c(), our
decision of whether or not we can possibly inline c()---depth and cost
limits notwithstanding---varies depending on whether we consider the
toplevel caller to be b() or c().

Since inlining decision-making no longer has access to the whole
callstack context, and instead just looks at a single caller-callee
pair, we need to detect these failures explicitly in RegionFormer::go()
and trigger retry.

Reviewed By: @ottoni

Differential Revision: D1452237